backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
psycopg2-binary
langchain-groq
langchain-huggingface
langchain-community
faiss-cpu
sentence-transformers
numpy
